Another tiny refactoring for #1318: make `Datacenter` a static inner class since it is not used by any other Model class.
