+++ restapi/src/main/java/io/stargate/web/restapi/resources/v2/schemas/KeyspacesResource.java
@@ -22,7 +22,6 @@ import io.stargate.auth.Scope;
 import io.stargate.auth.SourceAPI;
 import io.stargate.auth.entity.ResourceKind;
 import io.stargate.db.query.builder.Replication;
-import io.stargate.web.models.Datacenter;
 import io.stargate.web.models.Error;
 import io.stargate.web.models.Keyspace;
 import io.stargate.web.resources.Converters;
@@ -290,14 +289,14 @@ public class KeyspacesResource {
         });
   }
 
-  private List<Datacenter> buildDatacenters(io.stargate.db.schema.Keyspace keyspace) {
-    List<Datacenter> dcs = new ArrayList<>();
+  private List<Keyspace.Datacenter> buildDatacenters(io.stargate.db.schema.Keyspace keyspace) {
+    List<Keyspace.Datacenter> dcs = new ArrayList<>();
     for (Map.Entry<String, String> entries : keyspace.replication().entrySet()) {
       if (entries.getKey().equals("class") || entries.getKey().equals("replication_factor")) {
         continue;
       }
 
-      dcs.add(new Datacenter(entries.getKey(), Integer.parseInt(entries.getValue())));
+      dcs.add(new Keyspace.Datacenter(entries.getKey(), Integer.parseInt(entries.getValue())));
     }
 
     return dcs.isEmpty() ? null : dcs;
